Restructured setup managers&removed main.qml
