+++ PyTskMge/__main__.py
@@ -2,7 +2,7 @@ import sys
 from PySide6.QtWidgets import QMenuBar,QMainWindow,QApplication
 from PySide6.QtGui import QCloseEvent
 from taskmgm.taskmanagement import TaskManager
-from App import MenuSetup,LayoutSetup
+from App.Setups import SetupManager
 
 class QtApp:
     def __init__(self) -> None:
@@ -10,15 +10,14 @@ class QtApp:
         self.window = QMainWindow()
         self.taskmgr = TaskManager(self.window)
         self.window.closeEvent = self.close
+        self.sumgr = SetupManager(self,self.menubar)
     def setupMenuBar(self):
         self.window.resize(800,600)
         self.menubar = QMenuBar()
-        self.menusetup = MenuSetup.MenuSetupManager(self.menubar,self)
-        self.menusetup.taskmenu_setup()
+        self.sumgr.taskmenu_setup()
         self.window.setMenuBar(self.menubar)
     def setupLayout(self):
-        self.layoutsetup = LayoutSetup.LayoutSetupManager(self)
-        self.layoutsetup.mainLayout_setup()
+        self.sumgr.mainLayout_setup()
     def setupUI(self):
         self.setupMenuBar()
         self.setupLayout()
